+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next release
+
+### 🔧 Fixed
+
+* Do not attempt to reach the `lvmcryo` server when running in development mode.
+
+
 ## 0.2.21 - 2026-02-23
 
 ### ✨ Improved